Remove Float16 cast
I had the wrong selection, there was a String col in there. I don't
think it has a massive impact.

+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/fargate/estimate_ind_cqc_filled_posts_by_job_role.py
@@ -304,9 +304,7 @@ def main(
             pct_share_groups
         ).agg(
             metadata_selector.first(),
-            pl.struct(job_role_col, computed_cols.cast(pl.Float16)).alias(
-                "by_job_role_data"
-            ),
+            pl.struct(job_role_col, computed_cols).alias("by_job_role_data"),
             pl.len().alias("role_count"),
         )
